+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change log
 
+## v2.3
+- **2.3.0**
+  - add types
+  - remove just-mongo from main dependencies
+  - update just-mongo version and driver implementation
+  - minor fix of code style
+  - use default export (see the readme)
 ## v2.2
 - **2.2.1**
   - Make better redis driver